+++ packages/plugin-react-swc/src/index.ts
@@ -86,6 +86,7 @@ type Options = {
 const react = (_options?: Options): Plugin[] => {
   let hmrDisabled = false
   let base: string
+  let viteCacheRoot: string | undefined
   const options = {
     jsxImportSource: _options?.jsxImportSource ?? 'react',
     tsDecorators: _options?.tsDecorators,
@@ -141,6 +142,7 @@ const react = (_options?: Options): Plugin[] => {
       }),
       configResolved(config) {
         base = config.base
+        viteCacheRoot = config.cacheDir
         if (config.server.hmr === false) hmrDisabled = true
 
         const mdxIndex = config.plugins.findIndex(
@@ -192,6 +194,7 @@ const react = (_options?: Options): Plugin[] => {
           code,
           options.devTarget,
           options,
+          viteCacheRoot,
           {
             refresh,
             development: true,
@@ -219,11 +222,21 @@ const react = (_options?: Options): Plugin[] => {
           config: (userConfig) => ({
             build: silenceUseClientWarning(userConfig),
           }),
+          configResolved(config) {
+            viteCacheRoot = config.cacheDir
+          },
           transform: (code, _id) =>
-            transformWithOptions(_id.split('?')[0], code, 'esnext', options, {
-              runtime: 'automatic',
-              importSource: options.jsxImportSource,
-            }),
+            transformWithOptions(
+              _id.split('?')[0],
+              code,
+              'esnext',
+              options,
+              viteCacheRoot,
+              {
+                runtime: 'automatic',
+                importSource: options.jsxImportSource,
+              },
+            ),
         }
       : {
           name: 'vite:react-swc',
@@ -238,6 +251,9 @@ const react = (_options?: Options): Plugin[] => {
               },
             },
           }),
+          configResolved(config) {
+            viteCacheRoot = config.cacheDir
+          },
         },
   ]
 }
@@ -247,6 +263,7 @@ const transformWithOptions = async (
   code: string,
   target: JscTarget,
   options: Options,
+  viteCacheRoot: string | undefined,
   reactConfig: ReactConfig,
 ) => {
   const decorators = options?.tsDecorators ?? false
@@ -274,7 +291,10 @@ const transformWithOptions = async (
       jsc: {
         target,
         parser,
-        experimental: { plugins: options.plugins },
+        experimental: {
+          plugins: options.plugins,
+          cacheRoot: join(viteCacheRoot ?? 'node_modules/.vite', '.swc'),
+        },
         transform: {
           useDefineForClassFields: true,
           react: reactConfig,