+++ packages/plugin-rsc/examples/basic/src/routes/root.tsx
@@ -39,8 +39,13 @@ import { TestAssetsServer } from './assets/server'
 import { TestHmrSwitchServer } from './hmr-switch/server'
 import { TestHmrSwitchClient } from './hmr-switch/client'
 import { TestTreeShakeServer } from './tree-shake/server'
+import { TestTreeShake2 } from './tree-shake2/server'
 import { TestClientChunkServer } from './chunk/server'
 import { TestTailwind } from './tailwind'
+import { TestHmrClientDep2 } from './hmr-client-dep2/client'
+import { TestHmrClientDep3 } from './hmr-client-dep3/server'
+import { TestChunk2 } from './chunk2/server'
+import { TestUseId } from './use-id/server'
 
 export function Root(props: { url: URL }) {
   return (
@@ -63,7 +68,9 @@ export function Root(props: { url: URL }) {
         <TestTailwind />
         <TestDepCssInServer />
         <TestHydrationMismatch url={props.url} />
-        <TestHmrClientDep />
+        <TestHmrClientDep url={{ search: props.url.search }} />
+        <TestHmrClientDep2 url={{ search: props.url.search }} />
+        <TestHmrClientDep3 />
         <TestHmrSharedServer />
         <TestHmrSharedClient />
         <TestHmrSharedAtomic />
@@ -94,7 +101,10 @@ export function Root(props: { url: URL }) {
         <TestImportMetaGlob />
         <TestAssetsServer />
         <TestTreeShakeServer />
+        <TestTreeShake2 />
         <TestClientChunkServer />
+        <TestChunk2 />
+        <TestUseId />
       </body>
     </html>
   )