+++ packages/plugin-rsc/examples/basic/src/routes/tree-shake2/lib-client1.tsx
@@ -1,0 +1,5 @@
+'use client'
+
+export function LibClient1() {
+  return 'lib-client1'
+}