+++ packages/plugin-rsc/examples/basic/src/routes/tree-shake2/lib-server1.tsx
@@ -1,0 +1,3 @@
+export function LibServer1() {
+  return 'lib-server1'
+}