+++ packages/plugin-rsc/examples/basic/src/routes/tree-shake2/lib-server2.tsx
@@ -1,0 +1,3 @@
+export function LibServer2() {
+  return 'lib-server2:__unused_tree_shake2__'
+}