+++ packages/plugin-rsc/examples/basic/src/routes/tree-shake2/lib.tsx
@@ -1,0 +1,4 @@
+export * from './lib-client1'
+export * from './lib-client2'
+export * from './lib-server1'
+export * from './lib-server2'