+++ packages/plugin-rsc/src/index.ts
@@ -1,2 +1,6 @@
-export { default, type RscPluginOptions } from './plugin'
-export { transformHoistInlineDirective } from './transforms'
+export {
+  default,
+  type RscPluginOptions,
+  getPluginApi,
+  type PluginApi,
+} from './plugin'