+++ packages/plugin-rsc/src/transforms/fixtures/cjs/entry.mjs
@@ -1,3 +1,14 @@
 import depDefault from './dep1.cjs'
 import * as depNamespace from './dep2.cjs'
-export { depDefault, depNamespace }
+import depFn from './function.cjs'
+import depPrimitive from './primitive.cjs'
+import depExports from './exports.cjs'
+import depFnRequire from './function-require.cjs'
+export {
+  depDefault,
+  depNamespace,
+  depFn,
+  depPrimitive,
+  depExports,
+  depFnRequire,
+}